+++ src/onegov.core/onegov/core/orm/abstract/__init__.py
@@ -1,5 +1,10 @@
 from onegov.core.orm.abstract.adjacency_list import (
-    AdjacencyList, AdjacencyListCollection
+    AdjacencyList, AdjacencyListCollection, MoveDirection, sort_siblings
 )
 
-__all__ = ['AdjacencyList', 'AdjacencyListCollection']
+__all__ = [
+    'AdjacencyList',
+    'AdjacencyListCollection',
+    'MoveDirection',
+    'sort_siblings'
+]